ADOBE/1B/requirements.txt
PyMuPDF==1.23.8
scikit-learn==1.3.0
numpy==1.24.3
sentence-transformers==2.2.2
torch==2.0.1
joblib==1.3.2
pytesseract==0.3.10
Pillow==10.0.0 